requirements.txt
pyautogen
docker
packaging